agent/requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
aiosqlite==0.21.0
alembic==1.16.5
annotated-types==0.7.0
anthropic==0.67.0
anyio==4.9.0
appdirs==1.4.4
asgiref==3.9.1
asttokens==3.0.0
asyncio==3.4.3
attrs==25.3.0
Authlib==1.5.2
backoff==2.2.1
banks==2.2.0
bcrypt==4.3.0
beautifulsoup4==4.13.5
blinker==1.9.0
browserbase==1.4.0
build==1.3.0
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
chroma-hnswlib==0.7.6
chromadb==0.5.23
click==8.1.8
cloudpickle==3.1.1
cohere==5.18.0
colorama==0.4.6
coloredlogs==15.0.1
crewai==0.186.1
crewai-tools==0.71.0
cryptography==44.0.3
dataclasses-json==0.6.7
decorator==5.2.1
defusedxml==0.7.1
Deprecated==1.2.18
deprecation==2.1.0
dirtyjson==1.0.8
diskcache==5.6.3
distro==1.9.0
docker==7.1.0
docopt==0.6.2
docstring_parser==0.16
durationpy==0.10
embedchain==0.1.128
et_xmlfile==2.0.0
executing==2.2.1
fastapi==0.115.12
fastavro==1.12.0
ffmpeg==1.4
ffmpy==0.5.0
filelock==3.18.0
filetype==1.2.0
Flask==2.2.2
Flask-API==3.1
flask-healthz==0.0.3
flatbuffers==25.2.10
frozenlist==1.7.0
fsspec==2025.3.2
google-adk==0.4.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.169.0
google-auth==2.40.1
google-auth-httplib2==0.2.0
google-auth-oauthlib==0.8.0
google-cloud==0.34.0
google-cloud-aiplatform==1.92.0
google-cloud-appengine-logging==1.3.0
google-cloud-audit-log==0.2.5
google-cloud-bigquery==3.31.0
google-cloud-core==2.4.3
google-cloud-logging==3.5.0
google-cloud-resource-manager==1.14.2
google-cloud-secret-manager==2.23.3
google-cloud-speech==2.32.0
google-cloud-storage==2.19.0
google-cloud-texttospeech==2.27.0
google-cloud-trace==1.16.1
google-cloud-video-transcoder==1.15.1
google-crc32c==1.7.1
google-genai==1.14.0
google-generativeai==0.8.5
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
gptcache==0.1.44
gradio==5.29.1
gradio_client==1.10.1
graphviz==0.20.3
greenlet==3.2.4
griffe==1.14.0
groovy==0.1.2
grpc-google-iam-v1==0.14.2
grpcio==1.72.0rc1
grpcio-status==1.71.2
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.31.4
humanfriendly==10.0
hyperframe==6.1.0
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
importlib_metadata==8.6.1
importlib_resources==6.5.2
instructor==1.11.3
ipython==9.5.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.1.2
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.2
json5==0.12.1
json_repair==0.25.2
jsonpatch==1.33
jsonpickle==4.1.1
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kubernetes==33.1.0
lance-namespace==0.0.6
lance-namespace-urllib3-client==0.0.14
lancedb==0.25.0
langchain==0.3.27
langchain-cohere==0.3.5
langchain-community==0.3.29
langchain-core==0.3.76
langchain-experimental==0.3.4
langchain-openai==0.2.14
langchain-text-splitters==0.3.11
langgraph==0.6.7
langgraph-checkpoint==2.1.1
langgraph-prebuilt==0.6.4
langgraph-sdk==0.2.7
langsmith==0.3.45
litellm==1.74.9
llama-index-core==0.14.2
llama-index-instrumentation==0.4.1
llama-index-readers-file==0.5.4
llama-index-workflows==2.2.1
lxml==6.0.1
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib-inline==0.1.7
mcp==1.7.1
mdurl==0.1.2
mem0ai==0.1.117
mmh3==5.2.0
monotonic==1.6
moviepy==2.1.2
mpmath==1.3.0
multidict==6.6.4
mutagen==1.47.0
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.5
nltk==3.9.1
nodeenv==1.9.1
numpy==2.2.5
oauthlib==3.2.2
objprint==0.2.3
onnxruntime==1.22.0
openai==1.99.5
openpyxl==3.1.5
opentelemetry-api==1.32.1
opentelemetry-exporter-gcp-trace==1.9.0
opentelemetry-exporter-otlp-proto-common==1.32.1
opentelemetry-exporter-otlp-proto-grpc==1.32.1
opentelemetry-exporter-otlp-proto-http==1.32.1
opentelemetry-instrumentation==0.53b1
opentelemetry-instrumentation-asgi==0.53b1
opentelemetry-instrumentation-fastapi==0.53b1
opentelemetry-proto==1.32.1
opentelemetry-resourcedetector-gcp==1.9.0a0
opentelemetry-sdk==1.32.1
opentelemetry-semantic-conventions==0.53b1
opentelemetry-util-http==0.53b1
orjson==3.10.18
ormsgpack==1.10.0
overrides==7.7.0
packaging==25.0
pandas==2.2.3
parso==0.8.5
pdfminer.six==20250506
pdfplumber==0.11.7
pexpect==4.9.0
pillow==10.4.0
pipreqs==0.4.11
platformdirs==4.4.0
playwright==1.55.0
portalocker==2.7.0
posthog==3.25.0
proglog==0.1.12
prompt_toolkit==3.0.52
propcache==0.3.2
proto-plus==1.26.1
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.9
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydub==0.25.1
pyee==13.0.0
Pygments==2.19.1
PyJWT==2.10.1
pylance==0.36.0
pyparsing==3.2.3
pypdf==5.9.0
pypdfium2==4.30.0
PyPika==0.48.9
pyproject_hooks==1.2.0
pyright==1.1.405
pysbd==0.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-json-logger==2.0.7
python-multipart==0.0.20
pytube==15.0.0
pytz==2025.2
pyvis==0.3.2
PyYAML==6.0.2
qdrant-client==1.15.1
referencing==0.36.2
regex==2025.9.1
requests==2.32.5
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.27.1
rsa==4.9.1
ruff==0.11.10
safehttpx==0.1.6
schema==0.7.7
semantic-version==2.10.0
shapely==2.1.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.40
sse-starlette==2.3.4
stack-data==0.6.3
stagehand==0.5.2
starlette==0.46.2
striprtf==0.0.26
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.2
tiktoken==0.11.0
tinytag==2.1.1
tokenizers==0.20.3
tomli==2.2.1
tomli_w==1.2.0
tomlkit==0.13.2
tqdm==4.67.1
traitlets==5.14.3
typer==0.15.4
types-requests==2.32.4.20250913
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.1.1
urllib3==2.4.0
uv==0.8.17
uvicorn==0.34.2
uvloop==0.21.0
watchfiles==1.1.0
wcwidth==0.2.13
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==2.2.2
wrapt==1.17.2
xxhash==3.5.0
yarg==0.1.9
yarl==1.20.1
zipp==3.21.0
zstandard==0.23.0
